devfreq: Introduce devfreq boost driver
This driver boosts enumerated devfreq devices upon input, and allows for
boosting specific devfreq devices on other custom events. The boost
frequencies for this driver should be set so that frame drops are
near-zero at the boosted frequencies and power consumption is minimized
at said frequencies. The goal of this driver is to provide an interface
to achieve optimal device performance by requesting boosts on key
events, such as when a frame is ready to rendered to the display.

Currently, support is only present for boosting the cpu-llcc-ddr-bw
devfreq device, but the driver is structured in a way that makes it
easy to add support for new boostable devfreq devices in the future.

Change-Id: I2ea6fc2890d7267b3bb3474c0a2ae58ef4584057
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[Kazuki: Port to v5.4]
Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>
Signed-off-by: Sugakesshaa <sugakesshaa@gmail.com>